+++ lib/addic7ed.rb
@@ -8,6 +8,6 @@ require 'addic7ed/services/parse_page'
 require 'addic7ed/services/download_subtitle'
 require 'addic7ed/errors'
 require 'addic7ed/show_list'
-require 'addic7ed/video_file'
-require 'addic7ed/episode'
-require 'addic7ed/subtitle'
+require 'addic7ed/models/video_file'
+require 'addic7ed/models/episode'
+require 'addic7ed/models/subtitle'
