+++ lib/addic7ed/models/subtitle.rb
